pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "gendiff"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
repository = "https://github.com/zhukata/python-project-50.git"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
